fix for testing tm & dm accounts
